+++ R/methrix_object.R
@@ -13,6 +13,20 @@ methrix <- setClass(Class = 'methrix', contains = "SummarizedExperiment")
 # we can include validity checks as well
 # we might need to keep track of the processing steps
 
+#' extract CpGs per chromosome stats
+#' @name getChrSummary
+#' @rdname getChrSummary
+#' @param x An object of class methrix
+#' @return CpG per chromosome table
+#' @exportMethod getChrSummary
+#' @examples
+setGeneric(name = "getChrSummary", function(x) standardGeneric("getChrSummary"))
+
+## Accessor methods
+#' @rdname getChrSummary
+#' @aliases getChrSummary
+setMethod(f = "getChrSummary",signature = "methrix", function(x) x@metadata$chr_summary)
+
 setMethod(f = 'show', signature = 'methrix', definition = function(object){
   cat(paste('An object of class ', class(object), "\n"))
   print(object@metadata$summary)
@@ -22,12 +36,14 @@ setMethod(f = 'show', signature = 'methrix', definition = function(object){
 create_methrix = function(beta_mat = NULL, cov_mat = NULL, cpg_loci = NULL, is_hdf5 = FALSE,
                           genome_name = "hg19", col_data = NULL, h5_dir = NULL){
 
-  se_summary = data.table::data.table(ID = c("n_samples", "n_CpGs", "Reference_Build", "is_H5"),
-                                      Summary = c(ncol(beta_mat), nrow(beta_mat), genome_name, is_hdf5))
+  se_summary = data.table::data.table(ID = c("n_samples", "n_CpGs", "n_chromosomes", "Reference_Build", "is_H5"),
+                                      Summary = c(ncol(beta_mat), nrow(beta_mat), nrow(cpg_loci[,.N,chr]), genome_name, is_hdf5))
+
+  chr_summary = cpg_loci[,.N,chr]
 
   if(is_hdf5){
     se = SummarizedExperiment::SummarizedExperiment(assays = list(beta = as(beta_mat, "HDF5Array"), cov = as(cov_mat, "HDF5Array")),
-                                                    metadata = list(genome = genome_name, is_h5 = is_hdf5, summary = se_summary),
+                                                    metadata = list(genome = genome_name, is_h5 = is_hdf5, summary = se_summary, chr_summary = chr_summary),
                                                     colData = col_data, rowData = cpg_loci)
     if(!is.null(h5_dir)){
       tryCatch(HDF5Array::saveHDF5SummarizedExperiment(x = se, dir = h5_dir, replace = TRUE),
@@ -36,7 +52,7 @@ create_methrix = function(beta_mat = NULL, cov_mat = NULL, cpg_loci = NULL, is_h
   }else{
     se = SummarizedExperiment::SummarizedExperiment(assays = list(beta = data.table:::as.matrix.data.table(beta_mat),
                                                                   cov = data.table:::as.matrix.data.table(cov_mat)),
-                                                    metadata = list(genome = genome_name, is_h5 = is_hdf5, summary = se_summary),
+                                                    metadata = list(genome = genome_name, is_h5 = is_hdf5, summary = se_summary, chr_summary = chr_summary),
                                                     colData = col_data, rowData = cpg_loci)
   }